+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 1.0.0.alpha.3
 * **[FEATURE]** Added anonymous voting feature, toggleable in board admin panel (requested at [Feedbacky](https://app.feedbacky.net/i/anonymous-voting.730))
+* **[FEATURE]** Added administrator role that can do everything that owner can without ability to delete board
+(Partially implements request by [Matou0014](https://app.feedbacky.net/i/add-the-possibility-to-promote-users.2296))
 
 ### 1.0.0.alpha.2
 * **[IMPROVEMENT]** Added clearly visible Log In text in the log-in button